use npm workflow instead of homebrew actions
Against good practice this points against the master branches of the
action used, but given this is top tier players we may actually benefit
from this more than we'd see down sides.

This is using node 10 apparently but we "get a lot for free" so this
should not cause a real head ache.

See https://github.com/actions/npm