workflow: add release workflow
So far we tried having the publish step as a guarded step as part of the
default job run on push - with mixed success.
This now has redundancy and is something I don't particularly like.

Could be improved via:
* returning to a guarded step inside one job (many runs [6] on actual
release)
* having a separate publish job which 'needs' the build job inside the
same workflow (unclear how to have different 'on' for them)
* having multiple workflows but share actions for similar parts (still
lots of redundancy)